requirements.txt
django==4.1.7
django-oidc-user==0.3.6
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
django-cors-headers==3.13.0
pillow==9.3.0
python-dotenv==0.21.0
celery==5.2.7
redis==4.4.4
django-celery-results==2.4.0
django-storages==1.13.2
boto3==1.26.78
flower==1.2.0
